% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_helper.R
\name{add_mito_and_ribo}
\alias{add_mito_and_ribo}
\title{add_mito_and_ribo}
\usage{
add_mito_and_ribo(seurat, prefix = "percent.")
}
\arguments{
\item{prefix}{percent string prefix (optional; default: "percent.")}

\item{so}{Seurat object}
}
\value{
Seurat object with percentage mitochondrial reads and percent ribosomal reads
}
\description{
Adds percentage of mitochondrial reads and ribosomal reads for each cell
}
