% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monocle_wrappers.R
\name{add_pseudotime_metadata}
\alias{add_pseudotime_metadata}
\title{add_pseudotime_metadata}
\usage{
add_pseudotime_metadata(so, cds)
}
\arguments{
\item{so}{seurat object}

\item{cds}{cell data set object}
}
\value{
seurat object with pseudotime metadata
}
\description{
Adds pseudotime as metadata to seurat object
}
