% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_proportion.R
\name{calculate_proportion}
\alias{calculate_proportion}
\title{calculate_proportion}
\usage{
calculate_proportion(.data, var, norm_var)
}
\arguments{
\item{.data}{input data frame, where each row is an observation}

\item{var}{grouping var (for Seurat metadata, usually \code{clustering})}

\item{norm_var}{var to normalize against (for Seurat metadata, usually \code{orig.ident})}
}
\value{
tibble with proportion and normalized proportion
}
\description{
Calculate the proportion and normalized proportion of entries
represented within a given group
}
\details{
Example: metadata \%>\%
calculate_proportion(clustering, orig.ident) \%>\%
plot_prop(clustering, orig.ident)
}
