% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_helper.R
\name{get_gene_expr_from_so}
\alias{get_gene_expr_from_so}
\title{get_gene_expr_from_so}
\usage{
get_gene_expr_from_so(so, gene_vec, assay = "RNA", skip_missing = FALSE)
}
\arguments{
\item{so}{Seurat object}

\item{gene_vec}{character vector of gene names}

\item{assay}{assay to pull data from (optional; default:"RNA")}

\item{skip_missing}{bool, whether or not to skip missing genes. If true, will error}
}
\value{
tibble of gene expression per cell
}
\description{
Gets tibble of genes expression for select genes from Seurat object
}
