% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leading_edge_gene_analysis.R
\name{get_leading_edge_df}
\alias{get_leading_edge_df}
\title{get_leading_edge_df}
\usage{
get_leading_edge_df(rank_vec, gs_vec)
}
\arguments{
\item{rank_vec}{named vector of values}

\item{gs_vec}{vector of genes in gene set}
}
\value{
leading edge df, columns: name, value, in_gs, rank, running_es, direction, is_leading_edge
}
\description{
Creates a "leading edge df" for input into other leading edge
analysis functions
}
