% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_galaxy_plot}
\alias{make_galaxy_plot}
\title{make_galaxy_plot}
\usage{
make_galaxy_plot(
  metadata_df,
  sample_n = 800,
  umap1 = "UMAP_1",
  umap2 = "UMAP_2"
)
}
\arguments{
\item{metadata_df}{data frame with umap coordinates}

\item{sample_n}{number of cells to downsample to (optional; default: 800)}

\item{umap1}{string denoting UMAP_1 (optional; default: UMAP_1)}

\item{umap2}{string denoting UMAP_2 (optional; default: UMAP_2)}
}
\value{

}
\description{
Makes galaxy plot
}
