% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulk_create.R
\name{make_pseudobulk}
\alias{make_pseudobulk}
\title{make_pseudobulk}
\usage{
make_pseudobulk(so, split_var = NULL, sample_var, vars)
}
\arguments{
\item{so}{Seurat object}

\item{split_var}{as string, name of variable splitting pseudobulks (e.g. cluster)}

\item{sample_var}{as string, name of variable identifying samples}

\item{vars}{list of strings, name of variables of relevant conditions}
}
\value{
list of pseudobulks (i.e. matrices of aggregated counts)
}
\description{
Makes pseudobulk from Seurat object
}
