% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_umap_skeleton}
\alias{make_umap_skeleton}
\title{make_umap_skeleton}
\usage{
make_umap_skeleton(
  metadata_df,
  UMAP_1 = "UMAP_1",
  UMAP_2 = "UMAP_2",
  stroke_size = 0.1,
  pt_size = 0.5
)
}
\arguments{
\item{UMAP_1}{string denoting UMAP_1 (optional; default: UMAP_1)}

\item{UMAP_2}{string denoting UMAP_2 (optional; default: UMAP_2)}

\item{stroke_size}{stroke size (optional; default: 0.1)}

\item{pt_size}{point size (optional; default: 0.5)}

\item{.data}{data frame with umap coordinates}
}
\value{

}
\description{
Creates umap skeleton from data frame
}
