% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulk_helper.R
\name{pb_to_deseq}
\alias{pb_to_deseq}
\title{pb_to_deseq}
\usage{
pb_to_deseq(
  pb,
  sample_var,
  vars,
  formula_str = paste0("~", paste0(vars, collapse = "+"))
)
}
\arguments{
\item{pb}{mtx}

\item{sample_var}{as string, name of variable identifying samples}

\item{vars}{list of strings, name of variables of relevant conditions}
}
\value{
deseq object
}
\description{
pb to deseq
}
