% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{pca_gplot_scale}
\alias{pca_gplot_scale}
\title{pca_gplot_scale}
\usage{
pca_gplot_scale(matrix, sep, nameindex, dim1, dim2)
}
\arguments{
\item{matrix}{Data matrix with column names}

\item{sep}{Sep character in sample name}

\item{nameindex}{Part of name that you want to plot}

\item{dim1}{PC component on x-axis}

\item{dim2}{PC component on y-axis}
}
\value{
a ggplot object
}
\description{
Plots PCA given a matrix with column names of variables
}
\details{
Function given to me by Deb Sen, credit goes to Kevin Bi
}
