% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_helper.R
\name{pivot_FindAllMarkers_wider}
\alias{pivot_FindAllMarkers_wider}
\title{pivot_FindAllMarkers_wider}
\usage{
pivot_FindAllMarkers_wider(
  markers,
  pval = p_val_adj,
  pval_thresh = 0.05,
  log2FC = avg_log2FC,
  ident = cluster
)
}
\arguments{
\item{markers}{Seurat object}

\item{pval}{pval var (default: p_val_adj)}

\item{pval_thresh}{p value filtering threshold (default: 0.05)}

\item{log2FC}{log2 fold change var (default: avg_log2FC)}

\item{ident}{identity var (default: cluster)}
}
\value{
tibble of top markers per ident, all genes meeting p-value threshold, ordered by top log2FC
}
\description{
Takes output of \code{FindAllMarkers}, filters down to significant genes,
then builds a wide table of top markers for each ident specified
}
