% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leading_edge_gene_analysis.R
\name{plot_gsea_curve}
\alias{plot_gsea_curve}
\title{plot_gsea_curve}
\usage{
plot_gsea_curve(
  leading_edge_df,
  title,
  strwidth = 30,
  save = FALSE,
  output_dir = ""
)
}
\arguments{
\item{leading_edge_df}{dataframe output by get_leading_edge_df}

\item{title}{string; title of gene set}

\item{strwidth}{(optional; default 30) width of title strwrap}

\item{save}{(optional) TRUE to save the plot to the output directory}

\item{output_dir}{(optional) output directory if saving}
}
\value{
ggplot
}
\description{
Given a "leading edge df" and title,
plots GSEA curve with power = 0
}
