% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leading_edge_gene_analysis.R
\name{plot_gsea_curve_pretty}
\alias{plot_gsea_curve_pretty}
\title{plot_gsea_curve_pretty}
\usage{
plot_gsea_curve_pretty(leg_df, title = "", group_var, color_palette)
}
\arguments{
\item{title}{string; title of gene set}

\item{group_var}{grouping variable}

\item{color_palette}{vector of color characters}

\item{leading_edge_df}{dataframe output by get_leading_edge_df}
}
\value{
ggplot
}
\description{
Given a "leading edge df" and title,
plots GSEA curve with power = 0, but
prettier than \code{plot_gsea_curve}
}
