% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leading_edge_gene_analysis.R
\name{plot_gsea_leg_plots}
\alias{plot_gsea_leg_plots}
\title{plot_gsea_leg_plots}
\usage{
plot_gsea_leg_plots(leading_edge_gene_df, title = "", strwidth = 30, topn = 20)
}
\arguments{
\item{title}{string; title of gene set}

\item{strwidth}{(optional; default 30) width of title strwrap}

\item{topn}{(optional; default 20) the number of top leading edge genes to plot}

\item{leading_edge_df}{dataframe output by get_leading_edge_df}
}
\value{
grob of ggplots
}
\description{
Given a "leading edge df" and title,
plots GSEA curve with power = 0 and
plots a bar plot with the top leading edge genes
using \code{plot_grid}
}
