% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leading_edge_gene_analysis.R
\name{plot_top_leg}
\alias{plot_top_leg}
\title{plot_top_leg}
\usage{
plot_top_leg(
  leading_edge_gene_df,
  title,
  topn = 20,
  save = FALSE,
  output_dir = ""
)
}
\arguments{
\item{title}{string; title of gene set}

\item{topn}{(optional; default 20) the number of top leading edge genes to plot}

\item{save}{(optional) TRUE to save the plot to the output directory}

\item{output_dir}{(optional) output directory if saving}

\item{leading_edge_df}{dataframe output by get_leading_edge_df}
}
\value{
ggplot
}
\description{
Given a "leading edge df" and title,
plots a bar plot with the top leading edge genes
}
