% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monocle_wrappers.R
\name{run_monocle_trajectory}
\alias{run_monocle_trajectory}
\title{run_monocle_trajectory}
\usage{
run_monocle_trajectory(so, reduction_method = "UMAP")
}
\value{
cell data set object
}
\description{
Wrapper for the monocle pipeline:
so \%>\% as.cell_data_set() \%>\% cluster_cells() \%>\% learn_graph() \%>\% order_cells()
}
