% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP_helper.R
\name{scp_export_clustering_file}
\alias{scp_export_clustering_file}
\title{scp_export_clustering_file}
\usage{
scp_export_clustering_file(
  so,
  str_subset_regex = "RNA_snn_res.",
  output_dir,
  filename = "clustering_scp"
)
}
\arguments{
\item{so}{Seurat object}

\item{str_subset_regex}{regex string to select columns from Seurat object metadata (default: "RNA_snn_res.")}

\item{output_dir}{path to output directory}

\item{filename}{export filename (default: "clustering_scp")}
}
\value{
Seurat object with identity set to given ident
}
\description{
Create clustering file for export to the Single Cell Portal
https://singlecell.broadinstitute.org/single_cell
}
