% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_helper.R
\name{set_assay}
\alias{set_assay}
\title{set_assay}
\usage{
set_assay(so, assay)
}
\arguments{
\item{so}{Seurat object}

\item{assay}{Assay given as string}
}
\value{
Seurat object with default assay set to given assay
}
\description{
Sets default assay of Seurat object
Equivalent to DefaultAssay(so) <- "assay"
}
