% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_helper.R
\name{set_idents}
\alias{set_idents}
\title{set_idents}
\usage{
set_idents(so, assay)

set_idents(so, assay)
}
\arguments{
\item{so}{Seurat object}

\item{assay}{Assay given as string}

\item{ident}{Identity given as string}
}
\value{
Seurat object with identity set to given ident

Seurat object with default assay set to given assay
}
\description{
Sets identity of Seurat object
Equivalent to Ident(so) <- "ident"

Sets default assay of Seurat object
Equivalent to DefaultAssay(so) <- "assay"
}
