% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulk.R
\name{seurat_to_sce}
\alias{seurat_to_sce}
\title{seurat_to_sce}
\usage{
seurat_to_sce(so, split_var = NULL, sample_var, vars)
}
\arguments{
\item{so}{Seurat object}

\item{split_var}{as string, name of variable splitting pseudobulks (e.g. cluster)}

\item{sample_var}{as string, name of variable identifying samples}

\item{vars}{list of strings, name of variables of relevant condition}
}
\value{
sce object
}
\description{
Turns seurat objects into sce objects
}
