% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_proportion.R
\name{calculate_proportion}
\alias{calculate_proportion}
\title{calculate_proportion}
\usage{
calculate_proportion(.tbl, var1, var2)
}
\arguments{
\item{.tbl}{input data table}

\item{var1}{first grouping variable, typically \code{sample_ID}}

\item{var2}{second grouping variable}
}
\value{
tibble with proportions
}
\description{
calculate_proportion
}
