% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_binary_on_umap}
\alias{plot_binary_on_umap}
\title{plot_binary_on_umap}
\usage{
plot_binary_on_umap(
  metadata_df,
  binary_var,
  umap1 = "UMAP_1",
  umap2 = "UMAP_2"
)
}
\arguments{
\item{metadata_df}{data frame with umap coordinates and binary variable}

\item{binary_var}{binary variable, must be TRUE/FALSE}

\item{umap1}{string denoting UMAP_1 (optional; default: UMAP_1)}

\item{umap2}{string denoting UMAP_2 (optional; default: UMAP_2)}
}
\value{

}
\description{
Plots a binary variable on UMAP
}
