% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_helper.R
\name{set_idents}
\alias{set_idents}
\title{set_idents}
\usage{
set_idents(so, ident)
}
\arguments{
\item{so}{Seurat object}

\item{ident}{Identity given as string}
}
\value{
Seurat object with identity set to given ident
}
\description{
Sets identity of Seurat object
Equivalent to Ident(so) <- "ident"
}
